+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 Preparing new version: v1.0
 
 Added 'no_form_fields' option to adminTable class to avoid some fields manipulation through form (i.e insertion_date, last_edit_date, ...)
+Added loadController method to router class
 
 2012-03-21, abidibo  <abidibo@gmail.com>
 ---------------------------------------------